refactor: 파일명 및 API 엔드포인트 RESTful하게 수정
